XMLHttpRequest/tests/submissions/Opera/overridemimetype-headers-received-state-force-shiftjis.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: overrideMimeType() in HEADERS RECEIVED state, enforcing Shift-JIS encoding</title>
    <meta charset="utf-8">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-overridemimetype()-method" data-tested-assertations="/following::ol/li[1] /following::ol/li[4]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test();
      test.step(function() {
        var client = new XMLHttpRequest();
        client.onreadystatechange = function() {
          if(client.readyState===2)client.overrideMimeType('text/plain;charset=Shift-JIS');
          if (client.readyState !== 4) return;
          assert_equals( client.responseText, 'テスト', 'overrideMimeType() in HEADERS RECEIVED state set encoding' );
          test.done();
        };
        client.open("GET", "resources/content-type.php?content-type="+encodeURIComponent('text/html;charset=UTF-8')+'&content=%83%65%83%58%83%67');
        client.send( '' );
      });
    </script>
  </body>
</html>
